split Consume to a separate file
